bot/requirements.txt
python-telegram-bot==20.8
requests==2.31.0
redis==5.0.3